LPS-72818 Moves test
