+++ modules/apps/web-experience/site/site-test/src/testIntegration/java/com/liferay/site/service/persistence/test/GroupServiceTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.service;
+package com.liferay.site.service.persistence.test;
 
 import com.liferay.asset.kernel.service.AssetTagLocalServiceUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
